backend/requirements.txt
python-dotenv~=1.2.1
supabase~=2.22.3
fastapi~=0.120.2
pydantic~=2.12.3
passlib~=1.7.4
bcrypt==4.3.0
python-multipart~=0.0.9
